(#7) moved TestMode into separate file to reduce merge conflicts
